% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{Integrative function for building packages}
\usage{
buildPackage(
  package,
  version = NULL,
  path = pathGitHub(),
  register.c.routine = FALSE,
  compile.attributes = TRUE,
  update.collate = FALSE,
  update.date = TRUE,
  roxygenise = TRUE,
  build = TRUE,
  options.build = NULL,
  untar = TRUE,
  check = FALSE,
  options.check = NULL,
  install = TRUE,
  options.install = "--build",
  trace = 2,
  clear.existing = TRUE,
  clear.r.check = TRUE,
  clear.install = TRUE
)
}
\arguments{
\item{package}{the name of the package}

\item{version}{the version of the package}

\item{path}{the position of the directory containing the package}

\item{register.c.routine}{should the function \code{package_native_routine_registration_skeleton} be used to register C++ routines.}

\item{compile.attributes}{should the function \code{compile.attributes} be used to enerates the bindings required to call C++ functions from R for functions adorned with the Rcpp::export attribute.}

\item{update.collate}{should the collate field of the DESCRIPTION file be updated according the content of the R directory.}

\item{update.date}{should the date field of the DESCRIPTION file be updated with the date of the date.}

\item{roxygenise}{should the documentation be generated using \code{roxygenise}}

\item{build}{should the package be build}

\item{options.build}{additional options used to build the package}

\item{untar}{should the package be unzipped}

\item{check}{should the CRAN test be applied to the package}

\item{options.check}{additional options used to check the package}

\item{install}{should the package be installed on the computer.}

\item{options.install}{additional options used to install the package}

\item{trace}{How the execution of the function should be traced. If 2, also display the execution of compile.attributes.}

\item{clear.existing}{should the existing directories containing the tests, the unziped package or the archive file be removed.}

\item{clear.r.check}{Should the directory where the package has been tested be removed at the end of the execution.}

\item{clear.install}{Should the directory where the package has been unzipped be removed at the end of the execution.}
}
\description{
Convenient way to compile, test or install a package
}
\details{
Reminder [not related to the function]
exportPattern("^[^\\.]") : exporte toutes les fonctions ne commencant pas par un point
exportPattern("^[[:alpha:]]+") : exporte toutes les fonctions commencant par une lettre
}
\examples{
\dontrun{
buildPackage("butils")
}
}
