% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butils.base-package.R
\docType{package}
\name{butils.base}
\alias{butils.base-package}
\alias{butils.base}
\title{butils.base package}
\description{
Tools to source R code or install R packages
}
\author{
\strong{Maintainer}: Brice Ozenne \email{brice.mh.ozenne@gmail.com} (\href{https://orcid.org/0000-0001-9694-2956}{ORCID})

}
