% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declareRoutineC.R
\name{declareRoutineC}
\alias{declareRoutineC}
\title{Help function to fill the file declareRoutines.c [should be removed ?]}
\usage{
declareRoutineC(fct.name)
}
\arguments{
\item{fct.name}{A character string containig the name of the function}
}
\description{
Help function to fill the file declareRoutines.c
}
\examples{
\dontrun{
library(devtools)
library(tools)
path2package <- file.path(pathGitHub(),"lavaReduce")
setwd(path2package)
Rcpp::compileAttributes(pkgdir = path2package)
x <- capture.output(package_native_routine_registration_skeleton(dir = ".",,,FALSE))
cat(paste(x,collapse = "\\n"))


}

}
