% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencyPackage.R
\name{dependencyPackage}
\alias{dependencyPackage}
\title{Extract the Names of the Dependences of a Package}
\usage{
dependencyPackage(name, path, field = c("Imports", "Depends", "Suggests"))
}
\arguments{
\item{name}{[character] The name of the package.}

\item{path}{[character] The path to the directory containing the package.}

\item{field}{[character vector] which type of dependency should be considered?}
}
\value{
character vector containing the names of the packages
}
\description{
Extract the names of the dependences of a package.
}
\examples{
\dontrun{
  dependencyPackage("lavaSearch2", path = pathGitHub())
}

}
