% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIM.R
\name{DIM}
\alias{DIM}
\title{Dimensions of an Object}
\usage{
DIM(x)
}
\arguments{
\item{x}{object.}
}
\description{
Return \code{length} for vectors, lists, S4 objects, and use \code{dim} otherwise.
}
\details{
Used by emacs-config.
}
