% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRchunk.R
\name{extractRchunk}
\alias{extractRchunk}
\title{Extract code chunks from an .org file}
\usage{
extractRchunk(
  file,
  newfile = NULL,
  header = NULL,
  sep.header = NULL,
  sep.start.SRC = NULL,
  sep.end.SRC = NULL,
  rm.export.none = TRUE,
  rm.noexport = TRUE,
  index.chunk = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{file}{[character] name of the file from which the R chunks should be extracted.}

\item{newfile}{[character] name of the R file be created.}

\item{header}{[character vector] optional additional strings to be added at the beginning of the file.}

\item{sep.header}{[character] string indicating headers.}

\item{sep.start.SRC}{[character] string indicating the start of a code block.}

\item{sep.end.SRC}{[character] string indicating the end of a code block.}

\item{rm.export.none}{[logical] should chunks of code with \code{:export none} be ignored?}

\item{rm.noexport}{[logical]  should header with \code{:noexport:} be ignored?}

\item{index.chunk}{[logical]  should the code block be numbered in the .R file?}

\item{overwrite}{[logical] Can the existing R file be overwritten?}
}
\description{
Collect code chunks from an .org file and put them into an .R file.
}
