% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDocArgs.R
\name{extractSectionFromRD}
\alias{extractSectionFromRD}
\title{Extract sections from an .rd file}
\usage{
extractSectionFromRD(file)
}
\arguments{
\item{file}{the path to a .rd file}
}
\description{
Extract sections from an .rd file
}
\examples{

\dontrun{
path2packages <- file.path(pathGitHub(),"butils.base")
all.files <- list.files(file.path(path2packages,"man"), full.names = TRUE)
extractSectionFromRD(all.files[1])
}
}
