% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_SourceGitHub.R
\name{install_githubCommit}
\alias{install_githubCommit}
\title{Install commit}
\usage{
install_githubCommit(repo, ref = "master", subdir = NULL,
  host = "https://api.github.com", temporary = TRUE, force = FALSE,
  quiet = FALSE, ...)
}
\arguments{
\item{repo}{Repository address in the format username/repo}

\item{ref}{Desired git reference. See the documentation of \code{devtools::install_github}.}

\item{subdir}{Subdirectory within repo that contains the R package.  See the documentation of \code{devtools::install_github}.}

\item{host}{GitHub API host to use. See the documentation of \code{devtools::install_github}.}

\item{temporary}{Should the package be installed under a newname, i.e. adding \code{Tempo} to the name of the package}

\item{force}{Not sure what it is !!!}

\item{quiet}{Not sure what it is !!!}

\item{...}{additional arguments to be passed to \code{devtools::install_packages} or \code{devtools:::install}}
}
\description{
Install a specific commit of a package, possibly under a temporary name
}
\examples{
\dontrun{
commit <- read_githubCommit("bozenne/BuyseTest", seq.commit = 1)[1,"commits"]
install_githubCommit("bozenne/BuyseTest", ref = commit)
}

}
\keyword{function}
\keyword{github}
\keyword{package}

