% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDocArgs.R
\name{listDocArgs}
\alias{listDocArgs}
\title{Get all the arguments from the Rd files}
\usage{
listDocArgs(dir, trace = TRUE)
}
\arguments{
\item{dir}{the path to the directory containing the R package}

\item{trace}{should the execution of the function be traced}
}
\value{
a list containing:
\itemize{
\item argByFunction name of the arguments of the functions/methods that are documented.
\item docByFunction description of the arguments of the functions/methods that are documented.
\item functionNames name of the functions/methods that are documented.
\item dtAll data.table object gathering function names, argument names, and description of the arguments
}
}
\description{
Create a file in inst containing the description of all the arguments from the Rd file of a package
}
\examples{
\dontrun{
res <- listDocArgs(file.path(pathGitHub(),"butils.base"))
res$functionNames
res$argByFunction
}

}
