% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMES.R
\name{NAMES}
\alias{NAMES}
\title{Names of an Object}
\usage{
NAMES(x)
}
\arguments{
\item{x}{object.}
}
\description{
Return \code{colnames} for matrices, \code{isS4} for S4 objects, and return \code{names} otherwise.
}
\details{
Used by emacs-config.
}
