% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2script.R
\name{object2script}
\alias{object2script}
\alias{object2script.numeric}
\alias{object2script.character}
\alias{object2script.matrix}
\alias{object2script.data.frame}
\alias{object2script.data.table}
\alias{object2script.list}
\title{Display the Source Code Corresponding to an Object}
\usage{
object2script(object, ...)

\method{object2script}{numeric}(object, digits = NULL, print = TRUE, final.space = "\\n", ...)

\method{object2script}{character}(object, print = TRUE, final.space = "\\n", ...)

\method{object2script}{matrix}(object, digits = NULL, print = TRUE, final.space = "\\n", ...)

\method{object2script}{data.frame}(object, digits = NULL, print = TRUE, final.space = "\\n", ...)

\method{object2script}{data.table}(object, digits = NULL, print = TRUE, final.space = "\\n", ...)

\method{object2script}{list}(object, digits = NULL, print = TRUE, ...)
}
\arguments{
\item{object}{a vector, matrix, data.frame, data.table, or a list of these elements.}

\item{...}{for compatibility with the generic function.}

\item{digits}{[interger,>0] indicating the number of decimal places to be used. Passed to \code{round}.}

\item{print}{[logical] should the result be displayed.}

\item{final.space}{[character] Character string to be added at the end. For internal use.}
}
\description{
Display the source code corresponding to an object.
}
\details{
Used by emacs-config.
}
\examples{
## vector
vec <- 1:10
object2script(vec)
object2script(setNames(vec, 1:10))

## matrix
M <- matrix(rnorm(16), 4, 4)
object2script(M)

colnames(M) <- c("a","b","c","d")
object2script(M)

rownames(M) <- c("a","b","c","d")
object2script(M)

## data.frame
df <- data.frame(matrix(rnorm(16), 4, 4),"a")
object2script(df, digits = 5)

## data.table
dt <- data.table(matrix(rnorm(16), 4, 4),"a")
object2script(dt, digits = 5)

## list
l <- list(vec, dt = dt, df = df)
object2script(l)

}
