% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Source.R
\name{package.source}
\alias{package.source}
\title{Source a package directory}
\usage{
package.source(name, path = path_gitHub(), Rpackage = TRUE, Rcode = TRUE,
  RorderDescription = TRUE, attach = TRUE, Ccode = FALSE,
  rebuild = FALSE, warning = TRUE)
}
\arguments{
\item{name}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{Rpackage}{should the related R package be loaded}

\item{Rcode}{should all the .R be sourced}

\item{RorderDescription}{should the R files be sourced in the order indicate by collate}

\item{attach}{source the .onAttach function if it is present in the current environment}

\item{Ccode}{should all the .cpp file be source (using Rcpp::sourceCpp)}

\item{rebuild}{Force a rebuild of the shared library (from Rcpp:::sourceCpp).}

\item{warning}{should a warning be displayed if some of the R files are not sourced}
}
\description{
Source all the R and Cpp file contain in a package
}
\examples{
package.source("butils")
package.source("riskRegression")
}
\seealso{
\code{\link{path_gitHub}}
}

