% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Utils-Dir.R
\name{path_gitHub}
\alias{path_gitHub}
\title{Find GitHub directory}
\usage{
path_gitHub(user = NULL)
}
\arguments{
\item{user}{the name corresponding to the cession.}
}
\description{
Localise the github directory on linux or windows OS.
}
\examples{
path_gitHub()

}
\keyword{function}
\keyword{package}

