% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathGitHub.R
\name{pathGitHub}
\alias{pathGitHub}
\title{Find GitHub directory}
\usage{
pathGitHub(user = NULL)
}
\arguments{
\item{user}{the name corresponding to the cession.}
}
\description{
Localise the github directory on linux or windows OS.
}
\examples{
\dontrun{
pathGitHub()
}
}
\keyword{function}
\keyword{package}
