% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_PreparePackage.R
\name{printLsArgs}
\alias{printLsArgs}
\title{Get all the arguments from the Rd files}
\usage{
printLsArgs(dir, write.file = FALSE, filename = "List_of_arguments",
  path = NULL, trace = TRUE)
}
\arguments{
\item{dir}{the path to the directory containing the R package}

\item{write.file}{should the result be saved in a file?}

\item{filename}{the name of the file used to save the results.}

\item{path}{the path to the directory where the file should be created.}

\item{trace}{should the execution of the function be traced}
}
\description{
Create a file in inst containing the description of all the arguments from the Rd file of a package
}
