% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Utils-Description.R
\name{read_description}
\alias{read_description}
\title{Load description file}
\usage{
read_description(package, path, field = NULL, rmBlanck = TRUE,
  rmComma = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{field}{should a specific field of the description file be extracted}

\item{rmBlanck}{should the initial blanck be removed?}

\item{rmComma}{should any comma be removed?}
}
\description{
Extract the description file of a package
}
\examples{
\dontrun{
read_description <- butils:::read_description
read_description("butils", path = path_gitHub())
read_description("butils", path = path_gitHub(), field = "Collate")
read_description("butils", path = path_gitHub(), field = "Collate", rmBlanck = FALSE)

read_description("butils", path = path_gitHub(), field = "Version")
read_description("butils", path = path_gitHub(), field = "Imports")

read_description("riskRegression", path = path_gitHub(), field = "Collate")
read_description("lava.penalty", path = path_gitHub(), field = "Suggests")
}
}

