% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_SourceGitHub.R
\name{read_githubCommit}
\alias{read_githubCommit}
\title{Existing commits}
\usage{
read_githubCommit(repo, seq.commit = 1:5, rev.commit = FALSE,
  keep.author = TRUE, keep.name = TRUE, keep.time = TRUE, trace = TRUE,
  ...)
}
\arguments{
\item{repo}{Repository address in the format username/repo}

\item{seq.commit}{the sequence of commit to return}

\item{rev.commit}{shouidl the order of the commit be reverested (last becomes first).}

\item{keep.author}{should the author be returned}

\item{keep.name}{should the name of each commit be returned}

\item{keep.time}{should the time at which the commit was done be returned}

\item{trace}{should the progression in geting the name/time of the commit be displayed}

\item{...}{additional arguments to be passed to \code{read_html}.}
}
\description{
Get the existing commits corresponding to a repo on Github
}
\examples{
read_githubCommit("bozenne/BuyseTest")
read_githubCommit("bozenne/BuyseTest",rev.commit = TRUE)

}
\keyword{function}
\keyword{github}

