% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{readDescription}
\alias{readDescription}
\title{Load description file}
\usage{
readDescription(package, path, field = NULL, rm.blanck = TRUE,
  rm.comma = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{field}{should a specific field of the description file be extracted}

\item{rm.blanck}{should the initial blanck be removed?}

\item{rm.comma}{should any comma be removed?}
}
\description{
Extract the description file of a package
}
\examples{
\dontrun{
readDescription <- butils:::readDescription
readDescription("butils", path = pathGitHub())
readDescription("butils", path = pathGitHub(), field = "Collate")
readDescription("butils", path = pathGitHub(), field = "Collate", rm.blanck = FALSE)

readDescription("butils", path = pathGitHub(), field = "Version")
readDescription("butils", path = pathGitHub(), field = "Imports")

readDescription("riskRegression", path = pathGitHub(), field = "Collate")
readDescription("lava.penalty", path = pathGitHub(), field = "Suggests")
}
}
