% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-HTML.R
\name{readHtmlLinks}
\alias{readHtmlLinks}
\title{links of a url page}
\usage{
readHtmlLinks(x, ...)
}
\arguments{
\item{x}{the url page. Character.}

\item{...}{additional arguments to be passed to \code{read_html}.}
}
\description{
Get the links of a url page.
}
\details{
inspired from http://stackoverflow.com/questions/27297484/r-using-rvest-package-instead-of-xml-package-to-get-links-from-url
}
\examples{
readHtmlLinks("https://cran.r-project.org/")
readHtmlLinks("https://github.com/bozenne/BuyseTest")

}
\keyword{function}
\keyword{url}
