% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Debug.R
\name{revTraceback}
\alias{revTraceback}
\title{Get and Print Call Stacks (reverse order)}
\usage{
revTraceback(x = NULL, max.lines = getOption("deparse.max.lines"),
  trace = TRUE)
}
\arguments{
\item{x}{\code{NULL} (default, meaning \code{.Traceback}), or an
    integer count of calls to skip in the current stack, or a list or
    pairlist of deparsed calls.  See the details.}

\item{max.lines}{The maximum number of lines to be printed
    \emph{per call}.  The default is unlimited.}

\item{trace}{should the call stacks be output}
}
\description{
Same as \code{base::traceback} but in reverse order
}

