% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkDirectory.R
\name{sinkDirectory}
\alias{sinkDirectory}
\title{Import All .rds Files in a Directory.}
\usage{
sinkDirectory(
  path,
  string.keep = NULL,
  string.exclude = NULL,
  addMissingCol = FALSE,
  fixed = FALSE,
  trace = TRUE
)
}
\arguments{
\item{path}{[character] path to the directory.}

\item{string.keep}{[regexp] character string indicating files to import.}

\item{string.exclude}{[regexp] character string indicating files to ignore.}

\item{addMissingCol}{[logical] if a dataset does not have the same columns as
the other, the necessary empty columns are added to it with \code{NA} as values.}

\item{fixed}{[logical] If \code{TRUE}, \code{pattern} is a string to be matched as is.
Argument passed to \code{\link{grep}}.}

\item{trace}{[logical] should a progress bar be displayed tracking how many files
have been imported.}
}
\value{
A \code{data.table} object.
}
\description{
Import all .rds files in a directory.
}
\details{
The function first read the name of all the files in the directory.
Then if the argument \code{string.keep} is not \code{NULL}, it only retains the
files whose name contain \code{string.keep}.
Then if the argument \code{string.exclude} is not \code{NULL}, it excludes the
files whose name contain \code{string.exclude}.

Each file must contain a \code{data.table} object with the same columns,
so that they can be combined.
}
\author{
Brice Ozenne
}
