% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcePackage.R
\name{sourcePackage}
\alias{sourcePackage}
\title{Source a package directory}
\usage{
sourcePackage(name, path = pathGitHub(), r.package = TRUE, r.code = TRUE,
  r.order.collate = FALSE, onAttach = TRUE, onLoad = TRUE,
  c.code = FALSE, rebuild = FALSE, warning = TRUE)
}
\arguments{
\item{name}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{r.package}{should the related R package be loaded}

\item{r.code}{should all the .R be sourced}

\item{r.order.collate}{should the R files be sourced in the order indicate by collate}

\item{onAttach}{source the .onAttach function if it is present in the current environment}

\item{onLoad}{source the .onLoad function if it is present in the current environment}

\item{c.code}{should all the .cpp file be source (using Rcpp::sourceCpp)}

\item{rebuild}{Force a rebuild of the shared library (from Rcpp:::sourceCpp).}

\item{warning}{should a warning be displayed if some of the R files are not sourced}
}
\description{
Source all the R and Cpp file contain in a package
}
\examples{
package.source("butils")
package.source("riskRegression")
}
\seealso{
\code{\link{pathGitHub}}
}
