% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNIQUE.R
\name{UNIQUE}
\alias{UNIQUE}
\title{Unique Elements of an Object}
\usage{
UNIQUE(x)
}
\arguments{
\item{x}{object.}
}
\description{
Return \code{levels(droplevels(.))} for factor variables and otherwise \code{unique(.)}
}
