% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validFCTs}
\alias{validFCTs}
\alias{validCharacter}
\alias{validPath}
\title{Check argument in functions}
\usage{
validCharacter(
  value1,
  name1 = as.character(substitute(value1)),
  valid.length,
  valid.values = "character",
  refuse.NULL = TRUE,
  refuse.duplicates = FALSE,
  method = NULL,
  addPP = TRUE
)

validPath(
  value1,
  name1 = as.character(substitute(value1)),
  type,
  method = NULL,
  addPP = TRUE,
  extension = NULL,
  check.fsep = FALSE
)
}
\arguments{
\item{value1}{the value of the (first) argument to be checked}

\item{name1}{the name of the (first) argument.}

\item{valid.length}{the acceptable length(s) for the argument. If \code{NULL} no test is performed.}

\item{valid.values}{the acceptable value(s) for the argument. If \code{NULL} no test is performed. Can also be "character" or "character_or_logical".}

\item{refuse.NULL}{should an error be output if value is \code{NULL}.}

\item{refuse.duplicates}{should an error be output if value contains duplicated values.}

\item{method}{the name of the function using the argument.}

\item{addPP}{add ": " after the name of the function in the error message.}

\item{type}{Either "dir" or "file" to check whether to path points to an existing directory or file.}

\item{extension}{filter the files by the type of extention.}

\item{check.fsep}{display a warning when the separator is not correctly specified.}
}
\value{
An invisible \code{TRUE} or an error message.
}
\description{
Check the validity of the arguments in functions
}
\keyword{check}
\keyword{function}
