% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Utils-Description.R
\name{write_date}
\alias{write_date}
\title{write date field}
\usage{
write_date(package, path, trace = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{trace}{should the execution of the function be traced}
}
\description{
write the date field in the DESCRIPTION file of a package
}
\examples{
\dontrun{
write_date("butils", path = path_gitHub())
}
}

