% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_Utils-Description.R
\name{write_description}
\alias{write_description}
\title{Update description file}
\usage{
write_description(package, path, newfile, trace = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{newfile}{the content of the new description file}

\item{trace}{should the execution of the function be traced}
}
\description{
Overwrite the description file of a package
}
\examples{
\dontrun{
file <- read_description("butils", path = path_gitHub())
write_description("butils", path = path_gitHub(), file)
}
}
