% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-buildPackage.R
\name{writeDate}
\alias{writeDate}
\title{write date field}
\usage{
writeDate(package, path, trace = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{trace}{should the execution of the function be traced}
}
\description{
write the date field in the DESCRIPTION file of a package
}
\examples{
\dontrun{
writeDate("butils", path = pathGitHub())
}
}
