% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-buildPackage.R
\name{writeDescription}
\alias{writeDescription}
\title{Update description file}
\usage{
writeDescription(package, path, newfile, trace = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{path}{the path to the directory containing the package}

\item{newfile}{the content of the new description file}

\item{trace}{should the execution of the function be traced}
}
\description{
Overwrite the description file of a package
}
\examples{
\dontrun{
file <- readDescription("butils", path = pathGitHub())
writeDescription("butils", path = pathGitHub(), file)
}
}
