% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Dir.R
\name{cleanDir}
\alias{cleanDir}
\title{Remove directory}
\usage{
cleanDir(dir.name, test = TRUE, trace = TRUE)
}
\arguments{
\item{dir.name}{the path leading to the directory}

\item{test}{should the user be asked whether he really wants to remove the directory}

\item{trace}{should the execution of the function be traced}
}
\description{
Remove all files from a given directory
}
\examples{
\dontrun{
   dir.create("MyNewDirButils314")
   cleanDir("MyNewDirButils314", test = FALSE)
}

}
