% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcePackage.R
\name{sourcePackage}
\alias{sourcePackage}
\title{Source a package directory}
\usage{
sourcePackage(name, path = pathGitHub(), trace = FALSE,
  r.package = TRUE, field = c("Imports", "Depends"), r.code = TRUE,
  r.order.collate = FALSE, onAttach = TRUE, onLoad = TRUE,
  c.code = FALSE, rebuild = FALSE, warning = TRUE)
}
\arguments{
\item{name}{[character] The name of the package.}

\item{path}{[character] The path to the directory containing the package.}

\item{trace}{[logical]  Should the execution of the R and cpp files be traced?}

\item{r.package}{[logical] Should the related R package be loaded.}

\item{field}{[character vector] which type of dependency should be loaded?
Only active if \code{r.package} it \code{TRUE}.}

\item{r.code}{[logical] Should all the .R be sourced.}

\item{r.order.collate}{[logical] Should the R files be sourced in the order indicate by collate.}

\item{onAttach}{[logical] Source the .onAttach function if it is present in the current environment.}

\item{onLoad}{[logical] Source the .onLoad function if it is present in the current environment.}

\item{c.code}{[logical] Should all the .cpp file be source (using Rcpp::sourceCpp).}

\item{rebuild}{[logical] Force a rebuild of the shared library (from Rcpp:::sourceCpp).}

\item{warning}{[logical] Should a warning be displayed if some of the R files are not sourced.}
}
\description{
Source all the R and Cpp file contain in a package
}
\examples{
\dontrun{
sourcePackage("butils")
sourcePackage("riskRegression")
}

}
\seealso{
\code{\link{pathGitHub}}
}
